[-] Core : you need to cast to integer even after ip2long because it returns false for an IPv6

+++ classes/Customer.php
@@ -526,7 +526,7 @@ class CustomerCore extends ObjectModel
 		FROM `'._DB_PREFIX_.'customer` c
 		LEFT JOIN `'._DB_PREFIX_.'guest` g ON g.id_customer = c.id_customer
 		LEFT JOIN `'._DB_PREFIX_.'connections` co ON g.id_guest = co.id_guest
-		WHERE co.`ip_address` = \''.ip2long(trim($ip)).'\'');
+		WHERE co.`ip_address` = \''.(int)ip2long(trim($ip)).'\'');
 	}
 
 	/**
